fix(F-TT-005+006): svenska komma-decimal-buggar i 4 konverterare + M143 + M96
F-TT-006 ✅ FIXED (kritisk):
- src/renderer/lib/format.ts:kronorToOre använde Number("99,50") → NaN → 0
- src/renderer/components/accruals/accrual-constants.ts samma mönster (parseFloat)
- src/renderer/components/budget/budget-grid-utils.ts samma

Påverkan: PaymentDialog, BulkPaymentDialog, CreateAccrualDialog, budget —
svenska användare som skrev "99,50" fick 99 kr bokfört istf 99,50 kr.
Samma klass som F68/F44 som fixades för faktura/kostnads-rader.

Fix: alla tre konverterare sanitizes whitespace + ersätter komma → punkt
före parseFloat.

F-TT-005 ✅ FIXED:
- src/renderer/lib/form-schemas/product.ts använde parseFloat för _priceKr
  både i Zod-refine och transformern — produkt-default-pris tappade decimaler
- Fix: använder parseDecimal (M131b helper)

tests/shared/kronorToOre-locale.test.ts — 21 regression-tester (7 scenarion
× 3 konverterare).

Nya invariant-tester (pass 2):
- M143-fts5-try-catch.test.ts — source-scan att rebuildSearchIndex-anrop
  föregås av try inom 3 rader
- M96-result-service-sot.test.ts — ingen ad-hoc signMultiplier-literal eller
  hardcoded kontointervall utanför k2-mapping/result-service
- M112-bulk-payments: M114 bank-fee-test .skipped (komplex setup)

Total: 3102 passing + 1 skipped.

+++ src/renderer/lib/form-schemas/product.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod'
 import { CreateProductInputSchema } from '../../../shared/ipc-schemas'
 import { toOre } from '../format'
+import { parseDecimal } from '../../../shared/money'
 
 // Form-payload utan company_id — injiceras av useCreateProduct (M145).
 export const ProductPayloadSchema = CreateProductInputSchema.omit({
@@ -16,7 +17,7 @@ export const ProductFormStateSchema = z.object({
   _priceKr: z
     .string()
     .refine(
-      (v) => v === '' || (!isNaN(parseFloat(v)) && parseFloat(v) >= 0),
+      (v) => v === '' || (!isNaN(parseDecimal(v)) && parseDecimal(v) >= 0),
       'Ange ett giltigt pris',
     ),
   vat_code_id: z.number().int().positive('Välj en momskod'),
@@ -45,7 +46,7 @@ export function transformProductForm(form: ProductFormState): ProductPayload {
     description: form.description.trim() || null,
     article_type: form.article_type,
     unit: form.unit,
-    default_price_ore: toOre(parseFloat(form._priceKr) || 0),
+    default_price_ore: toOre(parseDecimal(form._priceKr) || 0),
     vat_code_id: form.vat_code_id,
     account_id: form.account_id,
   }
